requirements.txt
Flask==2.3.2
gunicorn==21.2.0
line-bot-sdk==3.18.1
openai>=1.0.0
setuptools>=65.0.0
wheel>=0.40.0
build>=1.0.3